Initial Commit - Almost fully implemented
The only instruction not implemented is opcode 6 which adds
two registers as if they were floating point numbers. I don't
need them and there's nothing else in the language or machine
that uses floating point numbers. If I need to implement it
later I can.

Otherwise, all opcodes are implemented.