Implemented opcodes D & E to allow a stack
Opcodes D & E allow using register values as
addresses to store and retrieve data. This
allows using a register as a stack pointer
to push and pop values on a stack. With that,
multi-depth function call chains can be implemented
with a link register and pushing link addresses
onto the stack before calling another function
and restoring the link function on return.