Removed legacy flag, swapped source and dest registers
The arithmetic instructions have the destination register
as the before the source registers. So the registers for
MOVR and STRR have been swapped to match. The first register
is the source of the move or store and the second is the
destination of the move or destination address of the store.