Fixed examples, updated readme, added small runtime
The runtime setups registers with common values and
sets the initial stack and link register values.
It provides and "exit" and "return" label. The exit
just runs HALT but can be used in the future for
any cleanup work. The return will jump to the address
stored in register F.

The runtime code (16 bytes) can be included by using
the compiler directive "@runtime". Directives must
be used before any other code.